requirements.txt
amqp==2.2.2
beautifulsoup4==4.5.0
billiard==3.5.0.3
celery==4.1.0
certifi==2018.1.18
chardet==3.0.4
Django==1.11.11
django-filter==1.1.0
djangorestframework==3.7.7
funcsigs==1.0.2
html5lib==1.0.1
idna==2.6
kombu==4.1.0
lxml==3.6.4
Markdown==2.6.11
mock==2.0.0
numpy==1.14.2
pandas==0.22.0
pbr==3.1.1
psycopg2==2.7.4
pyiso==0.4.0
python-dateutil==2.7.0
pytz==2018.3
requests==2.18.4
six==1.11.0
urllib3==1.22
vine==1.1.4
webencodings==0.5.1
xlrd==1.1.0
xmljson==0.1.9
